+++ test/reach/smell/checks/false_success_error_test.exs
@@ -20,6 +20,38 @@ defmodule Reach.Smell.Checks.FalseSuccessErrorTest do
     assert [%Finding{kind: :false_success_error}] = Smells.run(project)
   end
 
+  test "handles calls qualified through __MODULE__" do
+    project =
+      project_from_string(~S'''
+      defmodule Validator do
+        def validate_thing(arg) do
+          case __MODULE__.Helper.run(arg) do
+            {:ok, value} -> {:ok, value}
+            {:error, _reason} -> :ok
+          end
+        end
+      end
+      ''')
+
+    refute Enum.any?(Smells.run(project), &(&1.kind == :false_success_error))
+  end
+
+  test "detects suspicious module names qualified through __MODULE__" do
+    project =
+      project_from_string(~S'''
+      defmodule Validator do
+        def validate_thing(arg) do
+          case __MODULE__.Lint.run(arg) do
+            {:ok, value} -> value
+            {:error, _reason} -> :ok
+          end
+        end
+      end
+      ''')
+
+    assert [%Finding{kind: :false_success_error}] = Smells.run(project)
+  end
+
   test "allows error propagation" do
     project =
       project_from_string(~S'''